requirements.txt
matplotlib==3.9.4
numpy==1.26.4
pillow==11.3.0
protobuf==4.25.8
scipy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
torch==2.8.0
torchvision==0.23.0
websocket-client==1.9.0
